requirement.txt
# Backend libraries
numpy
pandas
seaborn
matplotlib
scikit-learn


# For chatbot
langchain
langchain-community  # newer LangChain plugins might require this
python-dotenv
transformers
langchain-huggingface  # if this is a real package; verify actual package name

# For CUDA 12.x support (PyTorch)
torch==2.1.0+cu121
torchvision==0.16.0+cu121
torchaudio==2.1.0+cu121
--extra-index-url https://download.pytorch.org/whl/cu121
#These torch liibraries according to my system , you may find dependencies according to your system on github 

# Frontend
streamlit
streamlit-option-menu
